requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.1
Faker==22.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
pillow==11.3.0
playwright==1.56.0
postgrest==2.25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.25.0
requests==2.32.5
six==1.17.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
websockets==15.0.1
yarl==1.22.0
